Show icon of package associated with Toast
This also fixes toast icon switch!

For all those times you have some random app or background service that posts
a Toast and you have no idea who's posting it.  This adds an icon badge to the
top left corner of the Toast to show the app's icon the Toast belongs to

Change-Id: I82bf23664eea134f3b1f89ad5a99f6be73906ba8
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>