OMS7-N: Support tagging resources as OK to overlay [1/11]
This will allow applications to have a resource xml defining what
resources that are safe to overlay by third party overlay packages.
The format of the tag is <overlay type="..." name="..."/> and it will
result in the FLAG_OVELAY being set on the resource entry.

An overlay package with resources that are not tagged as OK
to overlay by the target application, is considered to be a dangerous
overlay.

Idmaps generated for dangerous overlays will be flagged as dangerous in
the idmap header. It is still possible to use both idmap and overlays that
are dangerous, but it might not be advisable.

The intention is to allow dangerous overlays to be used if
they are pre-installed or if the signature of the overlay package
matches the signature of the target package.

Change-Id: I08d11de845c1679017798ea1636ef4c36f820d8e
(cherry picked from commit 7b1db662f29429470ae603f070aedbdb5851f155)